+++ CHANGELOG.md
@@ -10,6 +10,9 @@
         ([#4837](https://github.com/Automattic/pocket-casts-android/pull/4837))
     *   When following a podcast cache the smaller podcast images
         ([#4827](https://github.com/Automattic/pocket-casts-android/pull/4827))
+*   Bug Fixes
+    *   On Wear OS hide archived or played episodes
+        ([#4831](https://github.com/Automattic/pocket-casts-android/pull/4831))
 
 8.2
 -----